+++ src/heuristic/routing/ImplicitLoopDemandRoute.java
@@ -2,17 +2,17 @@ package heuristic.routing;
 
 public class ImplicitLoopDemandRoute extends DemandRoute<ImplicitLoopDemandRoute>{
 
-	public ImplicitLoopDemandRoute(double[][] distances, int[] demands) {
+	public ImplicitLoopDemandRoute(double[][] distances, double[] demands) {
 		super(distances, demands);
 	}
 
-	protected ImplicitLoopDemandRoute(double[][] distances, int[] demands,int[] stations,int size, double length, int demand) {
+	protected ImplicitLoopDemandRoute(double[][] distances, double[] demands,int[] stations,int size, double length, double demand) {
 		super(distances,demands,stations,size,length,demand);
 	}
 	
 	@Override
 	public void add(int station) {
-		int demandChange = demands[station];
+		double demandChange = demands[station];
 		double lengthChange;
 		
 		if(isEmpty()) {
@@ -31,7 +31,7 @@ public class ImplicitLoopDemandRoute extends DemandRoute<ImplicitLoopDemandRoute
 		if(n==0)
 			return;
 		
-		int demandChange = RoutingUtility.calculateDemand(n, stations, demands);
+		double demandChange = RoutingUtility.calculateDemand(n, stations, demands);
 		double lengthChange;
 		
 		if(isEmpty()) {
@@ -50,7 +50,7 @@ public class ImplicitLoopDemandRoute extends DemandRoute<ImplicitLoopDemandRoute
 	public int removeAt(int location) {
 		int rs = get(location);
 		
-		int demandChange = -1*demands[rs];
+		double demandChange = -1*demands[rs];
 		double lengthChange = 0;
 		
 		if(location==0) {
